Successful IDE HD read

+++ util.h
@@ -10,7 +10,7 @@ void memset(void *dest, uint8_t value, uint32_t byte_count);
 void outb(uint16_t port, uint8_t value);
 uint8_t inb(uint16_t port);
 void outw(uint32_t port, uint32_t value);
-uint32_t inw(uint32_t port);
+uint16_t inw(uint16_t port);
 void io_wait();
 
 #endif
